--- a/pkg/streamingpromql/testdata/ours/functions.test
+++ b/pkg/streamingpromql/testdata/ours/functions.test
@@ -970,3 +970,30 @@ eval instant at 1m stdvar_over_time(metric[1m1s])
   {case="Inf at start"} NaN
   {case="Inf at end"} NaN
   {case="all Inf"} NaN
+
+clear
+
+# Test edge cases for quantile_over_time.
+load 1m
+  metric{case="floats"} 1 2
+  metric{case="NaN at end"} 1 NaN
+  metric{case="NaN at start"} NaN 2
+  metric{case="all NaN"} NaN NaN
+  metric{case="Inf at start"} 1 Inf
+  metric{case="Inf at end"} Inf 2
+  metric{case="all Inf"} Inf Inf
+  metric{case="-Inf at start"} 1 -Inf
+  metric{case="-Inf at end"} -Inf 2
+  metric{case="all -Inf"} -Inf -Inf
+
+eval instant at 1m quantile_over_time(0.5, metric[1m1s])
+  {case="floats"} 1.5
+  {case="NaN at end"} NaN
+  {case="NaN at start"} NaN
+  {case="all NaN"} NaN
+  {case="Inf at start"} Inf
+  {case="Inf at end"} Inf
+  {case="all Inf"} Inf
+  {case="-Inf at start"} -Inf
+  {case="-Inf at end"} -Inf
+  {case="all -Inf"} -Inf