--- a/pkg/streamingpromql/testdata/upstream/functions.test
+++ b/pkg/streamingpromql/testdata/upstream/functions.test
@@ -1020,61 +1020,51 @@ load 10s
 	data_histogram{test="only histogram samples"} {{schema:0 sum:1 count:2}}x4
 	data_histogram{test="mix samples"} 0 1 2 {{schema:0 sum:1 count:2}}x2
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(0, data[2m])
-# 	{test="two samples"} 0
-# 	{test="three samples"} 0
-# 	{test="uneven samples"} 0
+eval instant at 1m quantile_over_time(0, data[2m])
+	{test="two samples"} 0
+	{test="three samples"} 0
+	{test="uneven samples"} 0
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(0.5, data[2m])
-# 	{test="two samples"} 0.5
-# 	{test="three samples"} 1
-# 	{test="uneven samples"} 1
+eval instant at 1m quantile_over_time(0.5, data[2m])
+	{test="two samples"} 0.5
+	{test="three samples"} 1
+	{test="uneven samples"} 1
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(0.75, data[2m])
-# 	{test="two samples"} 0.75
-# 	{test="three samples"} 1.5
-# 	{test="uneven samples"} 2.5
+eval instant at 1m quantile_over_time(0.75, data[2m])
+	{test="two samples"} 0.75
+	{test="three samples"} 1.5
+	{test="uneven samples"} 2.5
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(0.8, data[2m])
-# 	{test="two samples"} 0.8
-# 	{test="three samples"} 1.6
-# 	{test="uneven samples"} 2.8
+eval instant at 1m quantile_over_time(0.8, data[2m])
+	{test="two samples"} 0.8
+	{test="three samples"} 1.6
+	{test="uneven samples"} 2.8
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(1, data[2m])
-# 	{test="two samples"} 1
-# 	{test="three samples"} 2
-# 	{test="uneven samples"} 4
+eval instant at 1m quantile_over_time(1, data[2m])
+	{test="two samples"} 1
+	{test="three samples"} 2
+	{test="uneven samples"} 4
 
-# Unsupported by streaming engine.
-# eval_warn instant at 1m quantile_over_time(-1, data[2m])
-# 	{test="two samples"} -Inf
-# 	{test="three samples"} -Inf
-# 	{test="uneven samples"} -Inf
+eval_warn instant at 1m quantile_over_time(-1, data[2m])
+	{test="two samples"} -Inf
+	{test="three samples"} -Inf
+	{test="uneven samples"} -Inf
 
-# Unsupported by streaming engine.
-# eval_warn instant at 1m quantile_over_time(2, data[2m])
-# 	{test="two samples"} +Inf
-# 	{test="three samples"} +Inf
-# 	{test="uneven samples"} +Inf
+eval_warn instant at 1m quantile_over_time(2, data[2m])
+	{test="two samples"} +Inf
+	{test="three samples"} +Inf
+	{test="uneven samples"} +Inf
 
-# Unsupported by streaming engine.
-# eval_warn instant at 1m (quantile_over_time(2, (data[2m])))
-# 	{test="two samples"} +Inf
-# 	{test="three samples"} +Inf
-# 	{test="uneven samples"} +Inf
+eval_warn instant at 1m (quantile_over_time(2, (data[2m])))
+	{test="two samples"} +Inf
+	{test="three samples"} +Inf
+	{test="uneven samples"} +Inf
 
-# Unsupported by streaming engine.
-# eval instant at 1m quantile_over_time(0.5, data_histogram{test="only histogram samples"}[2m])
-# 	#empty
+eval instant at 1m quantile_over_time(0.5, data_histogram{test="only histogram samples"}[2m])
+	#empty
 
-# Unsupported by streaming engine.
-# eval_info instant at 1m quantile_over_time(0.5, data_histogram{test="mix samples"}[2m])
-# 	{test="mix samples"} 1
+eval_info instant at 1m quantile_over_time(0.5, data_histogram{test="mix samples"}[2m])
+	{test="mix samples"} 1
 
 clear